Refactor app into modular homeai package
